+++ OJS.Workers.SubmissionProcessors/LocalSubmissionsFilteringService.cs
@@ -1,0 +1,12 @@
+﻿namespace OJS.Workers.SubmissionProcessors
+{
+    using OJS.Workers.Common;
+    using OJS.Workers.SubmissionProcessors.Workers;
+
+    public class LocalSubmissionsFilteringService
+        : ISubmissionsFilteringService
+    {
+        public bool CanProcessSubmission(IOjsSubmission submission, ISubmissionWorker submissionWorker)
+            => submission != null;
+    }
+}
