+++ OJS.Workers.SubmissionProcessors/Models/TaskResultResponseModel.cs
@@ -4,11 +4,7 @@
 
     public class TaskResultResponseModel
     {
-        private const string ExecutionTimeValue = "just now";
-
         public int Points { get; set; }
-
-        public string TimeElapsedFormatted => ExecutionTimeValue;
 
         public IEnumerable<TestResultResponseModel> TestResults { get; set; }
     }
