--- OJS.Workers.SubmissionProcessors/OJS - Backup.Workers.SubmissionProcessors.csproj
@@ -1,53 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup Condition="'$(OS)' == 'Windows_NT'">
-    <TargetFrameworks>net47;netstandard2.0</TargetFrameworks>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(OS)' != 'Windows_NT'">
-    <TargetFramework>netstandard2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <Compile Remove="Models\RemoteSubmissionResult.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <None Include="Models\RemoteSubmissionResult.cs" />
-  </ItemGroup>
-
-  <!-- common NuGet package refs that affect all projects -->
-  <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.8" />
-  </ItemGroup>
-
-  <!-- .NET Standard 2.0 references, compilation flags and build options -->
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'netstandard2.0'">
-    <DefineConstants>NETCORE;NETSTANDARD;NETSTANDARD2_0</DefineConstants>
-  </PropertyGroup>
-
-  <!-- .NET 4.7 references, compilation flags and build options -->
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'net47'">
-    <DefineConstants>NET47;NETFRAMEWORK</DefineConstants>
-  </PropertyGroup>
-
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net47' ">
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Xml" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\OJS.Workers.Common\OJS.Workers.Common.csproj" />
-    <ProjectReference Include="..\OJS.Workers.ExecutionStrategies\OJS.Workers.ExecutionStrategies.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Folder Include="Workers\" />
-  </ItemGroup>
-</Project>
